requirements.txt
altgraph==0.17
future==0.18.2
greenlet==1.0.0
importlib-metadata==3.7.3
numpy==1.20.1
pandas==1.2.3
pefile==2019.4.18
pyaes==1.6.1
pyasn1==0.4.8
pyinstaller-hooks-contrib==2021.1
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
pywin32-ctypes==0.2.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.2
Telethon==1.21.1
typing-extensions==3.7.4.3
zipp==3.4.1
